refactor(richmond): material variant subfolders with shortened leaf names
Place named scene components under material folders (`floors/rough_stonework/arc_fill.rs`,
`partitions/rough_stonework/arc_15.rs`, `doors/wood/leaf.rs`, …) instead of
long flat filenames.

## Agent Dialogue

- **Developer:** Name the variant subfolder; e.g. `floors/rough_stonework/arc_fill.rs`; shorten partition leaf names too.
- **Agent:** Moved floors/roofs/stairs/doors into `rough_stonework/` and `wood/` subfolders; shortened partition modules to `arc_15`, `linear`, `header_15`, etc.
