+++ maybraid/richmond/building-components/src/doors.rs
@@ -1,14 +1,14 @@
 //! Door frame and leaf scene components.
 //!
-//! Pipeline: [`geometry`] → [`geometry_components`] → named door scene types.
+//! Pipeline: [`geometry`] → [`geometry_components`] → material variant modules.
 //! Frames expand into partition wall kit pieces.
 
 pub mod geometry;
 pub mod geometry_components;
 pub mod scene;
-pub mod wood_door_leaf;
+pub mod wood;
 
 pub use geometry::*;
 pub use geometry_components::DoorComponent;
 pub use scene::door_scene;
-pub use wood_door_leaf::WoodDoorLeaf;
+pub use wood::*;
