+++ maybraid/richmond/building-components/src/doors/wood.rs
@@ -1,0 +1,5 @@
+//! Wood door scene components.
+
+pub mod leaf;
+
+pub use leaf::WoodDoorLeaf;
