+++ maybraid/richmond/building-components/src/floors/rough_stonework.rs
@@ -1,0 +1,9 @@
+//! Rough stonework floor scene components.
+
+pub mod arc_fill;
+pub mod rectangle;
+pub mod struct_fill;
+
+pub use arc_fill::RoughStoneFloorArcFill;
+pub use rectangle::RoughStoneFloorRectangle;
+pub use struct_fill::RoughStoneFloorStructFill;
