+++ maybraid/richmond/building-components/src/floors/rough_stonework/arc_fill.rs
