+++ maybraid/richmond/building-components/src/floors/rough_stonework/rectangle.rs
