+++ maybraid/richmond/building-components/src/floors/rough_stonework/struct_fill.rs
