+++ maybraid/richmond/building-components/src/floors/wood.rs
@@ -1,0 +1,9 @@
+//! Wood floor scene components.
+
+pub mod arc_fill;
+pub mod rectangle;
+pub mod struct_fill;
+
+pub use arc_fill::WoodFloorArcFill;
+pub use rectangle::WoodFloorRectangle;
+pub use struct_fill::WoodFloorStructFill;
