+++ maybraid/richmond/building-components/src/floors/wood/rectangle.rs
