+++ maybraid/richmond/building-components/src/partitions/rough_stonework.rs
@@ -2,22 +2,22 @@
 //!
 //! Each type maps from a [`crate::partitions::geometry_components::WallComponent`] kit piece.
 
-pub mod rough_stonework_15;
-pub mod rough_stonework_180;
-pub mod rough_stonework_90;
-pub mod rough_stonework_header_15;
-pub mod rough_stonework_header_180;
-pub mod rough_stonework_header_90;
-pub mod rough_stonework_linear;
-pub mod rough_stonework_linear_header_subsegment;
-pub mod rough_stonework_linear_subsegment;
+pub mod arc_15;
+pub mod arc_180;
+pub mod arc_90;
+pub mod header_15;
+pub mod header_180;
+pub mod header_90;
+pub mod linear;
+pub mod linear_header_subsegment;
+pub mod linear_subsegment;
 
-pub use rough_stonework_15::RoughStonework15;
-pub use rough_stonework_180::RoughStonework180;
-pub use rough_stonework_90::RoughStonework90;
-pub use rough_stonework_header_15::RoughStoneworkHeader15;
-pub use rough_stonework_header_180::RoughStoneworkHeader180;
-pub use rough_stonework_header_90::RoughStoneworkHeader90;
-pub use rough_stonework_linear::RoughStoneworkLinear;
-pub use rough_stonework_linear_header_subsegment::RoughStoneworkLinearHeaderSubsegment;
-pub use rough_stonework_linear_subsegment::RoughStoneworkLinearSubsegment;
+pub use arc_15::RoughStonework15;
+pub use arc_180::RoughStonework180;
+pub use arc_90::RoughStonework90;
+pub use header_15::RoughStoneworkHeader15;
+pub use header_180::RoughStoneworkHeader180;
+pub use header_90::RoughStoneworkHeader90;
+pub use linear::RoughStoneworkLinear;
+pub use linear_header_subsegment::RoughStoneworkLinearHeaderSubsegment;
+pub use linear_subsegment::RoughStoneworkLinearSubsegment;
