+++ maybraid/richmond/building-components/src/partitions/rough_stonework/arc_15.rs
