+++ maybraid/richmond/building-components/src/partitions/rough_stonework/arc_180.rs
