+++ maybraid/richmond/building-components/src/partitions/rough_stonework/header_180.rs
