+++ maybraid/richmond/building-components/src/partitions/rough_stonework/linear.rs
