+++ maybraid/richmond/building-components/src/partitions/rough_stonework/linear_header_subsegment.rs
