+++ maybraid/richmond/building-components/src/partitions/rough_stonework/linear_subsegment.rs
