+++ maybraid/richmond/building-components/src/roofs.rs
@@ -1,17 +1,15 @@
 //! Roof / cap scene components.
 //!
-//! Pipeline: [`geometry`] → [`geometry_components`] → named roof scene types.
+//! Pipeline: [`geometry`] → [`geometry_components`] → material variant modules.
 
 pub mod geometry;
 pub mod geometry_components;
-pub mod rough_stone_perch_roof;
-pub mod rough_stone_spire_roof;
+pub mod rough_stonework;
 pub mod scene;
-pub mod wood_perch_deck;
+pub mod wood;
 
 pub use geometry::*;
 pub use geometry_components::RoofComponent;
-pub use rough_stone_perch_roof::RoughStonePerchRoof;
-pub use rough_stone_spire_roof::RoughStoneSpireRoof;
+pub use rough_stonework::*;
 pub use scene::roof_scene;
-pub use wood_perch_deck::WoodPerchDeck;
+pub use wood::*;
