+++ maybraid/richmond/building-components/src/roofs/rough_stonework.rs
@@ -1,0 +1,7 @@
+//! Rough stonework roof scene components.
+
+pub mod perch;
+pub mod spire;
+
+pub use perch::RoughStonePerchRoof;
+pub use spire::RoughStoneSpireRoof;
