+++ maybraid/richmond/building-components/src/roofs/rough_stonework/perch.rs
