+++ maybraid/richmond/building-components/src/roofs/rough_stonework/spire.rs
