+++ maybraid/richmond/building-components/src/roofs/wood.rs
@@ -1,0 +1,5 @@
+//! Wood roof / deck scene components.
+
+pub mod perch_deck;
+
+pub use perch_deck::WoodPerchDeck;
