+++ maybraid/richmond/building-components/src/stairs/rough_stonework.rs
@@ -1,0 +1,7 @@
+//! Rough stonework stair scene components.
+
+pub mod spiral;
+pub mod straight;
+
+pub use spiral::RoughStoneSpiralStair;
+pub use straight::RoughStoneStraightStair;
