+++ maybraid/richmond/building-components/src/stairs/rough_stonework/spiral.rs
