+++ maybraid/richmond/building-components/src/stairs/rough_stonework/straight.rs
