+++ maybraid/richmond/building-components/src/stairs/wood.rs
@@ -1,0 +1,5 @@
+//! Wood stair scene components.
+
+pub mod straight;
+
+pub use straight::WoodStraightStair;
