+++ maybraid/richmond/building-components/src/stairs/wood/straight.rs
